refactor 클래스 내 정의 순서 수정
