+++ src/main/java/domain/Cars.java
@@ -1,4 +1,4 @@
-package car;
+package domain;
 
 import utils.NumberGenerator;
