+++ src/main/java/domain/WinnerDeterminer.java
@@ -1,4 +1,4 @@
-package car;
+package domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
