+++ src/test/java/domain/CarTest.java
@@ -1,6 +1,5 @@
-package car;
+package domain;
 
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import utils.FixedNumberGenerator;
