+++ src/test/java/domain/WinnerDeterminerTest.java
@@ -1,4 +1,4 @@
-package car;
+package domain;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
